Создал кастомные классы исключений для использования в компоненте middleware, который отслеживает исключения на глобальном уровне. Данный компонент создал, чтобы из доменного слоя(наш сервис) мог производить валидацию модели и отправку(если необходимо) статусных кодов.Так же ввёл минимальную валидацию модели с помощью аттрибутов.

+++ ApplicationForVacancy/Service/Entity.cs
@@ -1,11 +1,9 @@
 ﻿using ApplicationForVacancy.Models;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Text;
 using System.Threading.Tasks;
 using ApplicationForVacancy.Context;
+using ApplicationForVacancy.CustomExceptionClasses;
 using Microsoft.EntityFrameworkCore;
 
 namespace ApplicationForVacancy.Service
@@ -19,20 +17,45 @@ namespace ApplicationForVacancy.Service
         }
         public async Task<Product> PostByNameAsync(string name)
         {
-            return await db.Product.FirstOrDefaultAsync(p=>p.Name==name);
+            var product = await db.Product.FirstOrDefaultAsync(p=>p.Name==name);
+            if (product == null)
+            {
+                throw new NotFoundException("Сущность не найдена");
+            }
+
+            return product;
         }
         public async Task<Product> GetByIdAsync(Guid id)
         {
-            return await db.Product.FirstOrDefaultAsync(p => p.Id==id);
+            var product = await db.Product.FirstOrDefaultAsync(p => p.Id==id);
+            if (product == null)
+            {
+                throw new NotFoundException("Сущность не найдена");
+            }
+
+            return product;
         }
         public async Task<Product> PostAddAsync(Product product)
         {
+            if (product == null)
+            {
+                throw new BadRequestException("Неверный запрос");
+            }
             db.Product.Add(product);
             await db.SaveChangesAsync();
             return product;
         }
         public async Task<Product> PutUpdAsync(Product product)
         {
+            if (product == null)
+            {
+                throw new BadRequestException("Неверный запрос");
+            }
+
+            if (!db.Product.Any(p => p.Id == product.Id))
+            {
+                throw new NotFoundException("Сущность не найдена");
+            }
             db.Update(product);
             await db.SaveChangesAsync();
             return product;
@@ -40,6 +63,10 @@ namespace ApplicationForVacancy.Service
         public async Task <Product> DeleteAsync(Guid id)
         {
             Product product = await db.Product.FirstOrDefaultAsync(p => p.Id==id);
+            if (product == null)
+            {
+                throw new NotFoundException("Сущность не найдена");
+            }
             db.Product.Remove(product);
             await db.SaveChangesAsync();
             return product;
